split corpus.rs into two different files
