+++ src/fuzzer/mod.rs
@@ -1,5 +1,6 @@
 pub mod cairoworker;
-pub mod corpus;
+pub mod corpus_crash;
+pub mod corpus_input;
 pub mod dict;
 pub mod fuzzer;
 pub mod starknetworker;
